Refactoring
refactoring based on https://canary.discord.com/channels/779549891890118683/994735343804883014/995345603636768808
