+++ src/main/java/mod/gate/core/events/ingamehud/ChatReceiveEvent.java
@@ -1,4 +1,4 @@
-package mod.gate.events.ingamehud;
+package mod.gate.core.events.ingamehud;
 
 import net.minecraft.network.MessageType;
 import net.minecraft.text.Text;
