+++ src/main/java/mod/gate/core/mixin/ClientChatReceiveMixin.java
@@ -1,7 +1,7 @@
-package mod.gate.mixin;
+package mod.gate.core.mixin;
 
-import mod.gate.events.ingamehud.ChatReceiveEvent;
-import mod.gate.events.EventHandler;
+import mod.gate.core.events.ingamehud.ChatReceiveEvent;
+import mod.gate.core.events.EventHandler;
 import net.minecraft.client.gui.hud.ChatHudListener;
 import net.minecraft.network.MessageType;
 import net.minecraft.text.Text;
