+++ src/main/java/mod/gate/features/NPCDialogue.java
@@ -5,8 +5,8 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import mod.gate.Gate;
-import mod.gate.events.ingamehud.ChatReceiveEvent;
-import mod.gate.events.Event;
+import mod.gate.core.events.ingamehud.ChatReceiveEvent;
+import mod.gate.core.events.Event;
 import mod.gate.utils.ChatUtils;
 import mod.gate.utils.Reference;
 import net.minecraft.client.MinecraftClient;
