+++ src/main/resources/gate.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "mod.gate.mixin",
+  "package": "mod.gate.core.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "ActionbarMixin",
